kitchen_recipe_chatbot_using_gemini_LLM_project/recipe_chatbot/requirements.txt
streamlit
google-generativeai
python-dotenv
nltk
textblob
black  # code formatter